+++ Changelog.md
@@ -14,6 +14,7 @@
  - Fix close button not working in the lobby for pairing 
  - Servers are now remembered after pairing with autoconnect support
  - Adds support for boundrys
+ - Adds support for microphone input
  
 # RC3